+++ zebra-chain/src/parameters/network_upgrade.rs
@@ -115,6 +115,7 @@ pub(super) const MAINNET_ACTIVATION_HEIGHTS: &[(block::Height, NetworkUpgrade)] 
         (NU6, Nu6),
         (NU6_1, Nu6_1),
         (NU6_2, Nu6_2),
+        (NU6_3, Nu6_3),
     ]
 };
 /// Testnet network upgrade activation heights.