+++ zebra-consensus/src/transaction/tests.rs
@@ -2504,7 +2504,7 @@ async fn v5_transaction_with_exceeding_expiry_height() {
         expiry_height,
         sapling_shielded_data: None,
         orchard_shielded_data: None,
-        network_upgrade: NetworkUpgrade::Nu6_2,
+        network_upgrade: NetworkUpgrade::Nu6_3,
     };
 
     let transaction_hash = transaction.hash();